LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/selectedcontent-nested-expected.txt



FAIL Modification behavior when nesting selectedcontent elements. assert_equals: child selectedcontent should not be modified after inserting. expected "child" but got "one"
